--- eth/backend_apis_mobile.go
@@ -1,9 +1,0 @@
-// +build mobile
-
-package eth
-
-import "github.com/ethereum/go-ethereum/rpc"
-
-func (s *Ethereum) APIs() []rpc.API {
-	return []rpc.API{}
-}
