+++ eth/tracers/tracers.go
@@ -1,4 +1,3 @@
-// +build !mobile
 // Copyright 2017 The go-ethereum Authors
 // This file is part of the go-ethereum library.
 //
